+++ sphinx/source/conf.py
@@ -228,8 +228,6 @@ man_pages = [
      'convert a skool file into a control file', _authors, 1),
     ('man/skool2html.py', 'skool2html.py',
      'convert a skool file and ref files to HTML', _authors, 1),
-    ('man/skool2sft.py', 'skool2sft.py',
-     'convert a skool file into a skool file template', _authors, 1),
     ('man/sna2ctl.py', 'sna2ctl.py',
      'generate a control file for a binary file or a snapshot', _authors, 1),
     ('man/sna2img.py', 'sna2img.py',